refactor: RouteChecker 삭제 후 Board에 원상복구
- 기존에 보드 클래스를 깔끔하게 하기 위해 뺐던 RouteChecker를 삭제하고 해당 로직을 다시 보드의 책임으로 원상복구
